requirements.txt
openai-agents==0.0.7
fastapi==0.115.12
uvicorn==0.34.0
asyncio==3.4.3
jinja2==3.1.6
requests==2.32.3
python-dotenv==1.1.0
fastmcp==0.4.1
